[TD-225]fix compiler error.

+++ src/os/src/detail/osString.c
@@ -58,7 +58,7 @@ bool taosMbsToUcs4(char *mbs, size_t mbsLength, char *ucs4, int32_t ucs4_max_len
 
   iconv_close(cd);
   if (len != NULL) {
-    *len = ucs4_max_len - outLeft;
+    *len = (int32_t)(ucs4_max_len - outLeft);
   }
 
   return true;
